Finished up match_scouting_auton and match_scouting_teleop. There is still a problem with the totalpoint counter - the app will crash once a cycle is scored. As such, right now it isn't shown on the app (It's commented out in the gridview children.)
Note - for the coopertition set / stack's points to be added to the currentcyclepoints counter at the top of the 
page, the tote button has to be pressed. (See ToteButton.Clicked.....for details)

Going to keep trying to fix the problem with the app crashing. Sorry to get this to you so late.